requirements.txt
Flask==3.0.0
requests==2.31.0
openai==1.54.0
python-dotenv==1.0.0
reportlab==4.0.7
Pillow==10.1.0
APScheduler==3.10.4
gunicorn==21.2.0
